+++ tests/unit/widgets/SocialShareTest.php
@@ -38,6 +38,7 @@ class SocialShareTest extends Unit
     private $tumblr = 'https://www.tumblr.com/widgets/share/tool?canonicalUrl=test+url&amp;posttype=link&amp;caption=test+title&amp;content=test+url';
     private $yahoo = 'https://compose.mail.yahoo.com/?subject=test+title&amp;body=test+description+-+test+url';
     private $odnoklassniki = 'https://connect.ok.ru/offer?url=test+url&amp;title=test+title&amp;description=test+description&amp;imageUrl=test+image+url';
+    private $trello = 'https://trello.com/add-card?url=test+url&amp;name=test+title';
 
     
     public function testInvalidConfigException()
@@ -65,6 +66,7 @@ class SocialShareTest extends Unit
                         . "<li><a href=\"$this->tumblr\" rel=\"noopener\" target=\"_blank\">Tumblr</a></li>"
                         . "<li><a href=\"$this->yahoo\" rel=\"noopener\" target=\"_blank\">Yahoo</a></li>"
                         . "<li><a href=\"$this->odnoklassniki\" rel=\"noopener\" target=\"_blank\">Odnoklassniki</a></li>"
+                        . "<li><a href=\"$this->trello\" rel=\"noopener\" target=\"_blank\">Trello</a></li>"
                     . "</ul>"
                     . "<!--/noindex-->";
         $actualHTML = $this->getActualHTML(self::DEFAULT_CONFIG_CONFIGURATOR_ID);
@@ -83,7 +85,7 @@ class SocialShareTest extends Unit
             $this->tester->metaTag('twitter:image', 'test image url'),
         ];
 
-        $this->assertEquals($expectedMetaTags, Yii::$app->getView()->metaTags);
+        $this->assertEquals($expectedMetaTags, array_values(Yii::$app->getView()->metaTags));
     }
 
     public function testDefaultIcons()
@@ -104,6 +106,7 @@ class SocialShareTest extends Unit
                 . "<li><a href=\"$this->tumblr\" rel=\"noopener\" target=\"_blank\"><i class=\"si si-tumblr\"></i></a></li>"
                 . "<li><a href=\"$this->yahoo\" rel=\"noopener\" target=\"_blank\"><i class=\"si si-yahoo\"></i></a></li>"
                 . "<li><a href=\"$this->odnoklassniki\" rel=\"noopener\" target=\"_blank\"><i class=\"si si-odnoklassniki\"></i></a></li>"
+                . "<li><a href=\"$this->trello\" rel=\"noopener\" target=\"_blank\"><i class=\"si si-trello\"></i></a></li>"
             . "</ul>"
             . "<!--/noindex-->";